Move wipe-database into database directory
