+++ index.js
@@ -8,7 +8,7 @@ const normalizeModel = require('./lib/resource-model/normalize');
 const buildJsonSchema = require('./lib/resource-definition/build-json-schema');
 const loadResourceModels = require('./lib/resource-model/load-from-disk');
 const validateResourceModel = require('./lib/resource-model/validate');
-const wipeDatabase = require('./lib/wipe-database');
+const wipeDatabase = require('./lib/database/wipe');
 
 // Options are all of the valid options for api-pls. Refer to the documentation
 // for the full list.
