+++ lib/database/wipe.js
